updated the workout lister to account for the user authorization

+++ frontend/src/components/WorkoutLister.js
@@ -1,13 +1,21 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
-
+import { useAuthContext } from "../hooks/useAuthContext";
 const WorkoutLister = ({ workout }) => {
 
+  const { user } = useAuthContext();
   const {dispatch} = useWorkoutsContext();
 
   const deleteWorkout = async () => {
 
+    if(!user){
+      return
+    }
+
     const response = await fetch(`/api/workouts/${workout._id}`,{
       method: "DELETE",
+      headers: {
+         'Authorization': `Bearer ${user.token}`
+      }
     });
 
     const json = await response.json();
